+++ src/channel.ts
@@ -63,6 +63,7 @@ import type {
   SearchPayload,
   SendMessageAPIResponse,
   SendMessageOptions,
+  SendReactionOptions,
   SharedLocationResponse,
   StaticLocationPayload,
   TruncateChannelAPIResponse,
@@ -423,7 +424,7 @@ export class Channel {
   async sendReaction(
     messageID: string,
     reaction: Reaction,
-    options?: { enforce_unique?: boolean; skip_push?: boolean },
+    options?: SendReactionOptions,
   ) {
     if (!messageID) {
       throw Error(`Message id is missing`);
@@ -467,7 +468,7 @@ export class Channel {
   async _sendReaction(
     messageID: string,
     reaction: Reaction,
-    options?: { enforce_unique?: boolean; skip_push?: boolean },
+    options?: SendReactionOptions,
   ) {
     if (!messageID) {
       throw Error(`Message id is missing`);