+++ src/types.ts
@@ -888,6 +888,7 @@ export type UserResponse = CustomUserData & {
   teams_role?: TeamsRole;
   updated_at?: string;
   username?: string;
+  avg_response_time?: number;
 };
 
 export type TeamsRole = { [team: string]: string };
@@ -2254,6 +2255,7 @@ export type AppSettings = {
   sqs_key?: string;
   sqs_secret?: string;
   sqs_url?: string;
+  user_response_time_enabled?: boolean;
   video_provider?: string;
   webhook_events?: Array<string> | null;
   webhook_url?: string;
@@ -2748,6 +2750,12 @@ export type SendMessageOptions = {
 
 export type UpdateMessageOptions = {
   skip_enrich_url?: boolean;
+  skip_push?: boolean;
+};
+
+export type SendReactionOptions = {
+  enforce_unique?: boolean;
+  skip_push?: boolean;
 };
 
 export type GetMessageOptions = {
@@ -2839,6 +2847,7 @@ export type Reaction = CustomReactionData & {
   score?: number;
   user?: UserResponse | null;
   user_id?: string;
+  emoji_code?: string;
 };
 
 export type Resource =